requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
nltk==3.8.1
pymongo==4.4.0
requests==2.31.0
beautifulsoup4==4.12.2
plotly==5.18.0
dash==2.11.0
vaderSentiment==3.3.2
python-dotenv==1.0.0
pyyaml==6.0
praw==7.8.0